requirements.txt
selenium
webdriver-manager
pandas
google-generativeai
python-dotenv
